backend/requirements.txt
flask==2.0.1
flask-cors==3.0.10
torch==2.0.0
torch-geometric==2.3.0
biopython==1.79
pandas==1.5.3
numpy==1.24.3
scikit-learn==1.2.2
pyjwt==2.3.0
werkzeug==2.0.3